+++ client/src/components/upload.jsx
@@ -6,13 +6,58 @@ import DropZone from './dropzone.jsx';
 class Upload extends Component {
   constructor(props) {
     super(props)
+    this.state = {
+      category: '',
+      location: '',
+      imageURL: '',
+      description: '',
+    };
+    this.getLink = this.getLink.bind(this);
+    this.handleInputChange = this.handleInputChange.bind(this);
+    this.handleSubmit = this.handleSubmit.bind(this);
+  }
+
+  getLink(imgurLink) {
+    this.setState({imageURL: imgurLink})
+  }
+
+  handleInputChange(event) {
+    const target = event.target;
+    const value = target.value;
+    const name = target.name;
+
+    this.setState({
+      [name]: value
+    });
+  }
+
+  handleSubmit(event) {
+    event.preventDefault();
+    console.log('submitted!');
+    axios.post(`/api/upload`, {
+      category: this.state.category,
+      location: this.state.location,
+      imageURL: this.state.imageURL,
+      description: this.state.description
+    })
+      .then(res => {
+        console.log(res);
+        console.log(res.data);
+      })
   }
 
   render() {
     return (<div>
       <Worthymap isForUploadPage={true} />
-      <DropZone />
-      <UploadForm />
+      <DropZone getLink={this.getLink}/>
+      <UploadForm 
+        category={this.state.category}
+        location={this.state.location}
+        imageURL={this.state.imageURL}
+        description={this.state.description}
+        handleInputChange={this.handleInputChange}
+        handleSubmit={this.handleSubmit}
+      />
     </div>)
   }
 }
