+++ client/src/components/worthymap.jsx
@@ -1,10 +1,6 @@
 import React, { Component } from 'react';
 import _ from 'lodash';
-import { compose, withProps, withHandlers, lifecycle } from 'recompose';
-import { withScriptjs, withGoogleMap, GoogleMap, Marker } from 'react-google-maps';
-import { Grid, Row } from 'react-bootstrap';
-import { MarkerClusterer }  from 'react-google-maps/lib/components/addons/MarkerClusterer';
-const { SearchBox } = require("react-google-maps/lib/components/places/SearchBox");
+import ClusteredMap from './clusteredmap.jsx';
 
 class WorthyMap extends Component {
   constructor(props) {
@@ -12,6 +8,8 @@ class WorthyMap extends Component {
     this.state = {
       markers: [],
       userMarker: [],
+      zoom: 4,
+      position: {lat: 37.09, lng: -95.71}
     };
     // zoom: props.zoom,
     // center: props.position
@@ -25,7 +23,7 @@ class WorthyMap extends Component {
 
   componentDidMount() {
     this.getUserLocation();
-    this.setState({markers: testingMarkers}) // delete this after getLocations is hooked up;
+    this.getLocations();
   }
 
   getUserLocation() {
@@ -55,9 +53,12 @@ class WorthyMap extends Component {
   }
 
   getLocations() {
-    // TODO
-    // Make axios request to our server to get all the locations
-    // turn them all into makers and call setState and set an array of markers to this.state.markers
+    if (!this.props.isForUploadPage) {
+      this.setState({markers: testingMarkers}) // delete this after getLocations is hooked up;
+      // TODO
+      // Make axios request to our server to get all the locations
+      // turn them all into makers and call setState and set an array of markers to this.state.markers
+    }
   }
 
   // allows a user to click the map to place a pin where there location is while uploading photos
@@ -81,7 +82,7 @@ class WorthyMap extends Component {
     } else {
       clickMap = () => {};
     }
-    console.log('renderprops', this.props.zoom, this.props.position)
+    console.log('renderprops', this.state.zoom, this.state.position)
     return (
       <ClusteredMap
         markers={ markers } 
@@ -93,128 +94,6 @@ class WorthyMap extends Component {
   }
 }
 
-// https://tomchentw.github.io/react-google-maps/#markerclusterer
-const ClusteredMap = compose(
-  withProps({
-    googleMapURL: "https://maps.googleapis.com/maps/api/js?key=AIzaSyC4R6AN7SmujjPUIGKdyao2Kqitzr1kiRg&v=3.exp&libraries=geometry,drawing,places",
-    loadingElement: <div style={{ height: `100%` }} />,
-    containerElement: <div style={{ height: `400px`, width: `100%`, border: `1px solid grey`, borderRadius: `3px`}}/>,
-    mapElement: <div style={{ height: `100%` }} />,
-  }),
-  withHandlers({
-    onMarkerClustererClick: () => (markerClusterer) => {
-      const clickedMarkers = markerClusterer.getMarkers()
-      console.log(`Current clicked markers length: ${clickedMarkers.length}`)
-      console.log(clickedMarkers)
-    },
-  }),
-  lifecycle({
-    componentWillMount() {
-      const refs = {}
-      this.setState({
-        zoom: null,
-        bounds: null,
-        center: null,
-        onMapMounted: ref => {
-          refs.map = ref;
-        },
-        onBoundsChanged: () => {
-          this.setState({
-            bounds: refs.map.getBounds(),
-            // center: refs.map.getCenter(),
-          })
-        },
-        onSearchBoxMounted: ref => {
-          refs.searchBox = ref;
-        },
-        onPlacesChanged: () => {
-          const places = refs.searchBox.getPlaces();
-          const bounds = new google.maps.LatLngBounds();
-
-          places.forEach(place => {
-            if (place.geometry.viewport) {
-              bounds.union(place.geometry.viewport)
-            } else {
-              bounds.extend(place.geometry.location)
-            }
-          });
-          const nextMarkers = places.map(place => ({
-            position: place.geometry.location,
-          }));
-          const nextCenter = _.get(nextMarkers, '0.position', this.state.center);
-
-          // hacky way of making sure the zoom updates even if the user changed the zoom
-          this.setState({
-            zoom: 1
-          })
-          this.setState({
-            center: nextCenter,
-            zoom: 11
-          });
-          console.log(this.state);
-        },
-      })
-    },
-  }),
-  withScriptjs,
-  withGoogleMap
-) ((props) => {
-  console.log('defaultzoom, defaultcenter, zoom, center', props.defaultZoom, props.defaultCenter, props.zoom, props.center)
-  const center = props.center || props.defaultCenter;
-  const zoom = props.zoom || props.defaultZoom;
-  const allMarkers = props.markers.map((marker, i) => (
-    <Marker
-      key={ i }
-      position={{ lat: marker.lat, lng: marker.lng }}
-      onClick={marker.clickHandler}
-    />
-  ));
-
-  return (
-    <GoogleMap
-      ref={props.onMapMounted}
-      zoom={ zoom }
-      center={ center }
-      onBoundsChanged={ props.onBoundsChanged }
-      onClick={ props.clickMap }
-    >
-      <MarkerClusterer
-        onClick={props.onMarkerClustererClick}
-        averageCenter
-        enableRetinaIcons
-        gridSize={60}
-      >
-        { allMarkers }
-      </MarkerClusterer>
-      <SearchBox
-        ref={props.onSearchBoxMounted}
-        bounds={props.bounds}
-        controlPosition={google.maps.ControlPosition.TOP_LEFT}
-        onPlacesChanged={props.onPlacesChanged}
-      >
-        <input
-          type="text"
-          placeholder="Search"
-          style={{
-            boxSizing: `border-box`,
-            border: `1px solid transparent`,
-            width: `240px`,
-            height: `32px`,
-            margin: 'auto auto auto auto',
-            marginTop: `10px`,
-            padding: `0 12px`,
-            borderRadius: `3px`,
-            boxShadow: `0 2px 6px rgba(0, 0, 0, 0.3)`,
-            fontSize: `14px`,
-            outline: `none`,
-            textOverflow: `ellipses`,
-            left: '500px'
-          }}
-        />
-      </SearchBox>
-    </GoogleMap>
-  );
-});
 
 const testingMarkers = [
   {
@@ -241,5 +120,6 @@ const testingMarkers = [
   }
 
 ]
+
 
 export default WorthyMap;
